+++ soql.js
@@ -1,17 +1,21 @@
 module.exports = function(RED) {
-    var nforce = require('./nforce_wrapper');
+    const nforce = require('./nforce_wrapper');
 
     function Query(config) {
         RED.nodes.createNode(this, config);
         this.connection = RED.nodes.getNode(config.connection);
-        var node = this;
+        const node = this;
+        var query = '';
         this.on('input', function(msg) {
             // show initial status of progress
             node.status({ fill: 'green', shape: 'ring', text: 'connecting....' });
 
             // use msg query if node's query is blank
+            // TODO: should the msg.query ALWAYS overwrite the config query of existent?
             if (msg.hasOwnProperty('query') && config.query === '') {
-                config.query = msg.query;
+                query = msg.query;
+            } else {
+                query = config.query;
             }
 
             // create connection object
@@ -21,7 +25,7 @@ module.exports = function(RED) {
             org
                 .authenticate({ username: this.connection.username, password: this.connection.password })
                 .then(function() {
-                    return org.query({ query: config.query });
+                    return org.query({ fetchAll: config.fetchAll, query: query });
                 })
                 .then(function(results) {
                     msg.payload = {
@@ -33,7 +37,7 @@ module.exports = function(RED) {
                 })
                 .error(function(err) {
                     node.status({ fill: 'red', shape: 'dot', text: 'Error!' });
-                    node.error(err);
+                    node.error(err, msg);
                 });
         });
     }